requirements.txt
PySastrawi==1.2.0
nltk==3.8.1
app-store-scraper==0.3.5
pandas==2.2.2
gensim==4.3.2
numpy==1.24.3
scipy==1.10.1
python-dotenv==1.0.1
google-cloud-bigquery==3.20.1
pyarrow==15.0.2
db-dtypes==1.2.0